+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Add `OpenTelemetry.Instrumentation.StackExchangeRedis` meter. Supported metrics:
+  * `redis.client.request.duration`,
+  * `redis.client.queue.duration`,
+  * `redis.client.network.duration`.
+  ([#1982](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1982)) 
+
 ## 1.9.0-beta.1
 
 Released 2024-Jul-23